+++ docs/changes.rst
@@ -1,15 +1,15 @@
 Sider Changelog
 ===============
 
-Branch ``sortedset``
---------------------
+Branch :branch:`sortedset`
+--------------------------
 
 - Added :mod:`sider.sortedset` module.
 - Added :class:`sider.types.SortedSet` type.
 
 
-Branch ``transaction``
-----------------------
+Branch :branch:`transaction`
+----------------------------
 
 Experiements about transactions.
 
@@ -18,11 +18,23 @@ Experiements about transactions.
 - Added :attr:`sider.session.Session.verbose_transaction_error` option.
 
 
-Version 0.1.2
+Version 0.1.3
 -------------
 
 To be released.
 
+- Now :class:`sider.hash.Hash` objects show their contents for :func:`repr()`.
+- Now persist objects show their key name for :func:`repr()`.
+- Added :data:`sider.lazyimport.exceptions` deferred module.
+
+
+Version 0.1.2
+-------------
+
+Released on April 11, 2012.  Pre-alpha release.
+
+- Now :class:`sider.session.Session` takes :class:`redis.client.StrictRedis`
+  object instead of :class:`redis.client.Redis` which is deprecated.
 - Added :mod:`sider.exceptions` module.
 - Added :class:`sider.warnings.SiderWarning` base class.
 - Fixed a bug of :meth:`sider.list.List.insert()` for index -1.